+++ app/controllers/folio/console/api/current_accounts_controller.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class Folio::Console::Api::CurrentAccountsController < Folio::Console::Api::BaseController
+  def console_path_ping
+    current_account.update_console_path!(params.require(:path))
+    head 200
+  end
+end